requirements.txt
python-telegram-bot==20.7
requests==2.31.0
sqlite3
configparser
google-generativeai==0.3.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3